@@ -2,7 +2,7 @@ name: flutter_goldens
 
 environment:
   # The pub client defaults to an <2.0.0 sdk constraint which we need to explicitly overwrite.
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.10.0-4.0.dev <2.10.0"
 
 dependencies:
   # To update these, use "flutter update-packages --force-upgrade".
