@@ -412,10 +412,11 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   final ValueNotifier<bool> _selectionEndInViewport = ValueNotifier<bool>(true);
 
   void _updateSelectionExtentsVisibility(Offset effectiveOffset) {
+    assert(selection != null);
     final Rect visibleRegion = Offset.zero & size;
 
     final Offset startOffset = _textPainter.getOffsetForCaret(
-      TextPosition(offset: _selection.start, affinity: _selection.affinity),
+      TextPosition(offset: selection.start, affinity: selection.affinity),
       _caretPrototype,
     );
     // TODO(justinmc): https://github.com/flutter/flutter/issues/31495
@@ -431,7 +432,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
       .contains(startOffset + effectiveOffset);
 
     final Offset endOffset =  _textPainter.getOffsetForCaret(
-      TextPosition(offset: _selection.end, affinity: _selection.affinity),
+      TextPosition(offset: selection.end, affinity: selection.affinity),
       _caretPrototype,
     );
     _selectionEndInViewport.value = visibleRegion
@@ -462,12 +463,8 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     // components, so always send those events, even if we didn't think it
     // changed. Also, focusing an empty field is sent as a selection change even
     // if the selection offset didn't change.
-    final bool focusingEmpty = nextSelection.baseOffset == 0
-      && nextSelection.extentOffset == 0
-      && !hasFocus;
-    if (nextSelection == selection
-        && cause != SelectionChangedCause.keyboard
-        && !focusingEmpty) {
+    final bool focusingEmpty = nextSelection.baseOffset == 0 && nextSelection.extentOffset == 0 && !hasFocus;
+    if (nextSelection == selection && cause != SelectionChangedCause.keyboard && !focusingEmpty) {
       return;
     }
     if (onSelectionChanged != null) {
@@ -514,7 +511,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   };
 
   void _handleKeyEvent(RawKeyEvent keyEvent) {
-    if(kIsWeb) {
+    if (kIsWeb) {
       // On web platform, we should ignore the key because it's processed already.
       return;
     }
@@ -533,6 +530,10 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
       // are pressed, just ignore the keypress.
       return;
     }
+
+    // TODO(ianh): It seems to be entirely possible for the selection to be null here, but
+    // all the keyboard handling functions assume it is not.
+    assert(selection != null);
 
     final bool isWordModifierPressed = isMacOS ? keyEvent.isAltPressed : keyEvent.isControlPressed;
     final bool isLineModifierPressed = isMacOS ? keyEvent.isMetaPressed : keyEvent.isAltPressed;
@@ -614,17 +615,18 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   }
 
   void _handleMovement(
-      LogicalKeyboardKey key, {
-      @required bool wordModifier,
-      @required bool lineModifier,
-      @required bool shift,
-    }) {
+    LogicalKeyboardKey key, {
+    @required bool wordModifier,
+    @required bool lineModifier,
+    @required bool shift,
+  }) {
     if (wordModifier && lineModifier) {
       // If both modifiers are down, nothing happens on any of the platforms.
       return;
     }
+    assert(selection != null);
 
-    TextSelection newSelection = selection;
+    TextSelection/*!*/ newSelection = selection;
 
     final bool rightArrow = key == LogicalKeyboardKey.arrowRight;
     final bool leftArrow = key == LogicalKeyboardKey.arrowLeft;
@@ -753,6 +755,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   // Handles shortcut functionality including cut, copy, paste and select all
   // using control/command + (X, C, V, A).
   Future<void> _handleShortcuts(LogicalKeyboardKey key) async {
+    assert(selection != null);
     assert(_shortcutKeys.contains(key), 'shortcut key $key not recognized.');
     if (key == LogicalKeyboardKey.keyC) {
       if (!selection.isCollapsed) {
@@ -802,6 +805,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   }
 
   void _handleDelete() {
+    assert(_selection != null);
     final String textAfter = selection.textAfter(_plainText);
     if (textAfter.isNotEmpty) {
       final int deleteCount = nextCharacter(0, textAfter);
@@ -839,7 +843,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
 
   // Returns a cached plain text version of the text in the painter.
   String _cachedPlainText;
-  String get _plainText {
+  String/*!*/ get _plainText {
     _cachedPlainText ??= _textPainter.text.toPlainText();
     return _cachedPlainText;
   }
@@ -847,7 +851,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   /// The text to display.
   TextSpan get text => _textPainter.text as TextSpan;
   final TextPainter _textPainter;
-  set text(TextSpan value) {
+  set text(TextSpan/*?*/ value) {
     if (_textPainter.text == value)
       return;
     _textPainter.text = value;
@@ -881,7 +885,10 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   /// its left.
   ///
   /// This must not be null.
-  TextDirection get textDirection => _textPainter.textDirection;
+  // TextPainter.textDirection is nullable, but it is set to a
+  // non-null value in the RenderEditable constructor and we refuse to
+  // set it to null here, so _textPainter.textDirection cannot be null.
+  TextDirection get textDirection => _textPainter.textDirection/*!*/;
   set textDirection(TextDirection value) {
     assert(value != null);
     if (_textPainter.textDirection == value)
@@ -1070,6 +1077,11 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   List<ui.TextBox> _selectionRects;
 
   /// The region of text that is selected, if any.
+  ///
+  /// The caret position is represented by a collapsed selection.
+  ///
+  /// If [selection] is null, there is no selection and attempts to
+  /// manipulate the selection will throw.
   TextSelection get selection => _selection;
   TextSelection _selection;
   set selection(TextSelection value) {
@@ -1231,13 +1243,22 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     markNeedsPaint();
   }
 
-  /// If false, [describeSemanticsConfiguration] will not set the
-  /// configuration's cursor motion or set selection callbacks.
+  /// Whether to allow the user to change the selection.
   ///
-  /// True by default.
+  /// Since [RenderEditable] does not handle selection manipulation
+  /// itself, this actually only affects whether the accessibility
+  /// hints provided to the system (via
+  /// [describeSemanticsConfiguration]) will enable selection
+  /// manipulation. It's the responsibility of this object's owner
+  /// to provide selection manipulation affordances.
+  ///
+  /// This field is used by [selectionEnabled] (which then controls
+  /// the accessibility hints mentioned above). When null,
+  /// [obscureText] is used to determine the value of
+  /// [selectionEnabled] instead.
   bool get enableInteractiveSelection => _enableInteractiveSelection;
   bool _enableInteractiveSelection;
-  set enableInteractiveSelection(bool value) {
+  set enableInteractiveSelection(bool/*?*/ value) {
     if (_enableInteractiveSelection == value)
       return;
     _enableInteractiveSelection = value;
@@ -1245,18 +1266,26 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     markNeedsSemanticsUpdate();
   }
 
-  /// {@template flutter.rendering.editable.selectionEnabled}
-  /// True if interactive selection is enabled based on the values of
+  /// Whether interactive selection are enabled based on the values of
   /// [enableInteractiveSelection] and [obscureText].
   ///
-  /// By default [enableInteractiveSelection] is null, obscureText is false,
-  /// and this method returns true.
-  /// If [enableInteractiveSelection] is null and obscureText is true, then this
-  /// method returns false. This is the common case for password fields.
-  /// If [enableInteractiveSelection] is non-null then its value is returned. An
-  /// app might set it to true to enable interactive selection for a password
-  /// field, or to false to unconditionally disable interactive selection.
-  /// {@endtemplate}
+  /// Since [RenderEditable] does not handle selection manipulation
+  /// itself, this actually only affects whether the accessibility
+  /// hints provided to the system (via
+  /// [describeSemanticsConfiguration]) will enable selection
+  /// manipulation. It's the responsibility of this object's owner
+  /// to provide selection manipulation affordances.
+  ///
+  /// By default, [enableInteractiveSelection] is null, [obscureText] is false,
+  /// and this getter returns true.
+  ///
+  /// If [enableInteractiveSelection] is null and [obscureText] is true, then this
+  /// getter returns false. This is the common case for password fields.
+  ///
+  /// If [enableInteractiveSelection] is non-null then its value is
+  /// returned. An application might [enableInteractiveSelection] to
+  /// true to enable interactive selection for a password field, or to
+  /// false to unconditionally disable interactive selection.
   bool get selectionEnabled {
     return enableInteractiveSelection ?? !obscureText;
   }
@@ -1264,9 +1293,12 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   /// The color used to paint the prompt rectangle.
   ///
   /// The prompt rectangle will only be requested on non-web iOS applications.
-  Color get promptRectColor => _promptRectPaint.color;
-  set promptRectColor(Color newValue) {
-    // Painter.color can not be null.
+  // TODO(ianh): We should change the getter to return null when _promptRectRange is null
+  // (otherwise, if you set it to null and then get it, you get back non-null).
+  // Alternatively, we could stop supporting setting this to null.
+  Color/*?*/ get promptRectColor => _promptRectPaint.color;
+  set promptRectColor(Color/*?*/ newValue) {
+    // Painter.color cannot be null.
     if (newValue == null) {
       setPromptRectRange(null);
       return;
@@ -1287,7 +1319,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   /// The prompt rectangle will only be requested on non-web iOS applications.
   ///
   /// When set to null, the currently displayed prompt rectangle (if any) will be dismissed.
-  void setPromptRectRange(TextRange newRange) {
+  void setPromptRectRange(TextRange/*?*/ newRange) {
     if (_promptRectRange == newRange)
       return;
 
@@ -1301,7 +1333,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   /// text is entered or removed in order to accommodate expanding when
   /// [expands] is set to true.
   double get maxScrollExtent => _maxScrollExtent;
-  double _maxScrollExtent = 0;
+  double/*!*/ _maxScrollExtent = 0;
 
   double get _caretMargin => _kCaretGap + cursorWidth;
 
@@ -1337,14 +1369,14 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     if (hasFocus && selectionEnabled)
       config.onSetSelection = _handleSetSelection;
 
-    if (selectionEnabled && _selection?.isValid == true) {
-      config.textSelection = _selection;
-      if (_textPainter.getOffsetBefore(_selection.extentOffset) != null) {
+    if (selectionEnabled && selection?.isValid == true) {
+      config.textSelection = selection;
+      if (_textPainter.getOffsetBefore(selection.extentOffset) != null) {
         config
           ..onMoveCursorBackwardByWord = _handleMoveCursorBackwardByWord
           ..onMoveCursorBackwardByCharacter = _handleMoveCursorBackwardByCharacter;
       }
-      if (_textPainter.getOffsetAfter(_selection.extentOffset) != null) {
+      if (_textPainter.getOffsetAfter(selection.extentOffset) != null) {
         config
           ..onMoveCursorForwardByWord = _handleMoveCursorForwardByWord
           ..onMoveCursorForwardByCharacter = _handleMoveCursorForwardByCharacter;
@@ -1352,38 +1384,45 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     }
   }
 
+  // TODO(ianh): in theory, [selection] could become null between when
+  // we last called describeSemanticsConfiguration and when the
+  // callbacks are invoked, in which case the callbacks will crash...
+
   void _handleSetSelection(TextSelection selection) {
     _handleSelectionChange(selection, SelectionChangedCause.keyboard);
   }
 
   void _handleMoveCursorForwardByCharacter(bool extentSelection) {
-    final int extentOffset = _textPainter.getOffsetAfter(_selection.extentOffset);
+    assert(selection != null);
+    final int extentOffset = _textPainter.getOffsetAfter(selection.extentOffset);
     if (extentOffset == null)
       return;
-    final int baseOffset = !extentSelection ? extentOffset : _selection.baseOffset;
+    final int baseOffset = !extentSelection ? extentOffset : selection.baseOffset;
     _handleSelectionChange(
       TextSelection(baseOffset: baseOffset, extentOffset: extentOffset), SelectionChangedCause.keyboard,
     );
   }
 
   void _handleMoveCursorBackwardByCharacter(bool extentSelection) {
-    final int extentOffset = _textPainter.getOffsetBefore(_selection.extentOffset);
+    assert(selection != null);
+    final int extentOffset = _textPainter.getOffsetBefore(selection.extentOffset);
     if (extentOffset == null)
       return;
-    final int baseOffset = !extentSelection ? extentOffset : _selection.baseOffset;
+    final int baseOffset = !extentSelection ? extentOffset : selection.baseOffset;
     _handleSelectionChange(
       TextSelection(baseOffset: baseOffset, extentOffset: extentOffset), SelectionChangedCause.keyboard,
     );
   }
 
   void _handleMoveCursorForwardByWord(bool extentSelection) {
-    final TextRange currentWord = _textPainter.getWordBoundary(_selection.extent);
+    assert(selection != null);
+    final TextRange currentWord = _textPainter.getWordBoundary(selection.extent);
     if (currentWord == null)
       return;
     final TextRange nextWord = _getNextWord(currentWord.end);
     if (nextWord == null)
       return;
-    final int baseOffset = extentSelection ? _selection.baseOffset : nextWord.start;
+    final int baseOffset = extentSelection ? selection.baseOffset : nextWord.start;
     _handleSelectionChange(
       TextSelection(
         baseOffset: baseOffset,
@@ -1394,13 +1433,14 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   }
 
   void _handleMoveCursorBackwardByWord(bool extentSelection) {
-    final TextRange currentWord = _textPainter.getWordBoundary(_selection.extent);
+    assert(selection != null);
+    final TextRange currentWord = _textPainter.getWordBoundary(selection.extent);
     if (currentWord == null)
       return;
     final TextRange previousWord = _getPreviousWord(currentWord.start - 1);
     if (previousWord == null)
       return;
-    final int baseOffset = extentSelection ?  _selection.baseOffset : previousWord.start;
+    final int baseOffset = extentSelection ?  selection.baseOffset : previousWord.start;
     _handleSelectionChange(
       TextSelection(
         baseOffset: baseOffset,
@@ -1441,7 +1481,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   // TODO(jonahwilliams): replace when we expose this ICU information.
   bool _onlyWhitespace(TextRange range) {
     for (int i = range.start; i < range.end; i++) {
-      final int codeUnit = text.codeUnitAt(i);
+      final int/*!*/ codeUnit = text.codeUnitAt(i);
       if (!_isWhitespace(codeUnit)) {
         return false;
       }
@@ -1475,7 +1515,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
 
   Axis get _viewportAxis => _isMultiline ? Axis.vertical : Axis.horizontal;
 
-  Offset get _paintOffset {
+  Offset/*!*/ get _paintOffset {
     switch (_viewportAxis) {
       case Axis.horizontal:
         return Offset(-offset.pixels, 0.0);
@@ -1485,7 +1525,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     return null;
   }
 
-  double get _viewportExtent {
+  double/*!*/ get _viewportExtent {
     assert(hasSize);
     switch (_viewportAxis) {
       case Axis.horizontal:
@@ -1496,7 +1536,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     return null;
   }
 
-  double _getMaxScrollExtent(Size contentSize) {
+  double/*!*/ _getMaxScrollExtent(Size contentSize) {
     assert(hasSize);
     switch (_viewportAxis) {
       case Axis.horizontal:
@@ -1866,8 +1906,6 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     return TextSelection(baseOffset: line.start, extentOffset: line.end);
   }
 
-  Rect _caretPrototype;
-
   void _layoutText({ double minWidth = 0.0, double maxWidth = double.infinity }) {
     assert(maxWidth != null && minWidth != null);
     if (_textLayoutLastMaxWidth == maxWidth && _textLayoutLastMinWidth == minWidth)
@@ -1884,6 +1922,8 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     _textLayoutLastMaxWidth = maxWidth;
   }
 
+  /*late*/ Rect _caretPrototype;
+
   // TODO(garyq): This is no longer producing the highest-fidelity caret
   // heights for Android, especially when non-alphabetic languages
   // are involved. The current implementation overrides the height set
@@ -1895,25 +1935,27 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   /// On iOS, the cursor is taller than the cursor on Android. The height
   /// of the cursor for iOS is approximate and obtained through an eyeball
   /// comparison.
-  Rect get _getCaretPrototype {
+  void _computeCaretPrototype() {
     assert(defaultTargetPlatform != null);
     switch (defaultTargetPlatform) {
       case TargetPlatform.iOS:
       case TargetPlatform.macOS:
-        return Rect.fromLTWH(0.0, 0.0, cursorWidth, cursorHeight + 2);
+        _caretPrototype = Rect.fromLTWH(0.0, 0.0, cursorWidth, cursorHeight + 2);
+        break;
       case TargetPlatform.android:
       case TargetPlatform.fuchsia:
       case TargetPlatform.linux:
       case TargetPlatform.windows:
-        return Rect.fromLTWH(0.0, _kCaretHeightOffset, cursorWidth, cursorHeight - 2.0 * _kCaretHeightOffset);
+        _caretPrototype = Rect.fromLTWH(0.0, _kCaretHeightOffset, cursorWidth, cursorHeight - 2.0 * _kCaretHeightOffset);
+        break;
     }
-    return null;
   }
+
   @override
   void performLayout() {
     final BoxConstraints constraints = this.constraints;
     _layoutText(minWidth: constraints.minWidth, maxWidth: constraints.maxWidth);
-    _caretPrototype = _getCaretPrototype;
+    _computeCaretPrototype();
     _selectionRects = null;
     // We grab _textPainter.size here because assigning to `size` on the next
     // line will trigger us to validate our intrinsic sizes, which will change
@@ -1949,6 +1991,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     assert(_textLayoutLastMaxWidth == constraints.maxWidth &&
            _textLayoutLastMinWidth == constraints.minWidth,
       'Last width ($_textLayoutLastMinWidth, $_textLayoutLastMaxWidth) not the same as max width constraint (${constraints.minWidth}, ${constraints.maxWidth}).');
+    assert(_caretPrototype != null);
 
     // If the floating cursor is enabled, the text cursor's color is [backgroundCursorColor] while
     // the floating cursor's color is _cursorColor;
@@ -1979,7 +2022,7 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
         case TargetPlatform.windows:
           // Override the height to take the full height of the glyph at the TextPosition
           // when not on iOS. iOS has special handling that creates a taller caret.
-          // TODO(garyq): See the TODO for _getCaretPrototype.
+          // TODO(garyq): See the TODO for _computeCaretPrototype().
           caretRect = Rect.fromLTWH(
             caretRect.left,
             caretRect.top - _kCaretHeightOffset,
@@ -2038,8 +2081,8 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     // We always want the floating cursor to render at full opacity.
     final Paint paint = Paint()..color = _cursorColor.withOpacity(0.75);
 
-    double sizeAdjustmentX = _kFloatingCaretSizeIncrease.dx;
-    double sizeAdjustmentY = _kFloatingCaretSizeIncrease.dy;
+    double/*!*/ sizeAdjustmentX = _kFloatingCaretSizeIncrease.dx;
+    double/*!*/ sizeAdjustmentY = _kFloatingCaretSizeIncrease.dy;
 
     if (_resetFloatingCursorAnimationValue != null) {
       sizeAdjustmentX = ui.lerpDouble(sizeAdjustmentX, 0, _resetFloatingCursorAnimationValue);
@@ -2155,16 +2198,17 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     bool showSelection = false;
     bool showCaret = false;
 
-    if (_selection != null && !_floatingCursorOn) {
-      if (_selection.isCollapsed && _showCursor.value && cursorColor != null)
+    if (selection != null && !_floatingCursorOn) {
+      if (selection.isCollapsed && _showCursor.value && cursorColor != null)
         showCaret = true;
-      else if (!_selection.isCollapsed && _selectionColor != null)
+      else if (!selection.isCollapsed && _selectionColor != null)
         showSelection = true;
       _updateSelectionExtentsVisibility(effectiveOffset);
     }
 
     if (showSelection) {
-      _selectionRects ??= _textPainter.getBoxesForSelection(_selection, boxHeightStyle: _selectionHeightStyle, boxWidthStyle: _selectionWidthStyle);
+      assert(selection != null);
+      _selectionRects ??= _textPainter.getBoxesForSelection(selection, boxHeightStyle: _selectionHeightStyle, boxWidthStyle: _selectionWidthStyle);
       _paintSelection(context.canvas, effectiveOffset);
     }
 
@@ -2175,8 +2219,10 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
     if (paintCursorAboveText)
       _textPainter.paint(context.canvas, effectiveOffset);
 
-    if (showCaret)
-      _paintCaret(context.canvas, effectiveOffset, _selection.extent);
+    if (showCaret) {
+      assert(selection != null);
+      _paintCaret(context.canvas, effectiveOffset, selection.extent);
+    }
 
     if (!paintCursorAboveText)
       _textPainter.paint(context.canvas, effectiveOffset);
@@ -2243,10 +2289,11 @@ class RenderEditable extends RenderBox with RelayoutWhenSystemFontsChangeMixin {
   @override
   List<DiagnosticsNode> debugDescribeChildren() {
     return <DiagnosticsNode>[
-      text.toDiagnosticsNode(
-        name: 'text',
-        style: DiagnosticsTreeStyle.transition,
-      ),
+      if (text != null)
+        text.toDiagnosticsNode(
+          name: 'text',
+          style: DiagnosticsTreeStyle.transition,
+        ),
     ];
   }
 }
