@@ -4,11 +4,12 @@
 
 import 'package:meta/meta.dart';
 import 'package:vm_snapshot_analysis/treemap.dart';
+import 'package:archive/archive.dart';
+import 'package:archive/archive_io.dart';
 
 import '../base/file_system.dart';
 import '../convert.dart';
 import 'logger.dart';
-import 'process.dart';
 import 'terminal.dart';
 
 /// A class to analyze APK and AOT snapshot and generate a breakdown of the data.
@@ -16,13 +17,11 @@ class SizeAnalyzer {
   SizeAnalyzer({
     @required this.fileSystem,
     @required this.logger,
-    @required this.processUtils,
     this.appFilenamePattern = 'libapp.so',
   });
 
   final FileSystem fileSystem;
   final Logger logger;
-  final ProcessUtils processUtils;
   final Pattern appFilenamePattern;
   String _appFilename;
 
@@ -30,44 +29,24 @@ class SizeAnalyzer {
 
   static const int tableWidth = 80;
 
-  /// Analyzes [apk] and [aotSnapshot] to output a [Map] object that includes
-  /// the breakdown of the both files, where the breakdown of [aotSnapshot] is placed
-  /// under 'lib/arm64-v8a/$_appFilename'.
-  ///
-  /// The [aotSnapshot] can be either instruction sizes snapshot or v8 snapshot.
-  Future<Map<String, dynamic>> analyzeApkSizeAndAotSnapshot({
-    @required File apk,
+  static const int _kAotSizeMaxDepth = 2;
+  static const int _kZipSizeMaxDepth = 1;
+
+  /// Analyze the [aotSnapshot] in an uncompressed output directory.
+  Future<Map<String, dynamic>> analyzeAotSnapshot({
+    @required Directory outputDirectory,
     @required File aotSnapshot,
+    @required File precompilerTrace,
+    @required String type,
+    String excludePath,
   }) async {
     logger.printStatus('▒' * tableWidth);
-    _printEntitySize(
-      '${apk.basename} (total compressed)',
-      byteSize: apk.lengthSync(),
-      level: 0,
-      showColor: false,
-    );
     logger.printStatus('━' * tableWidth);
-    final Directory tempApkContent = fileSystem.systemTempDirectory.createTempSync('flutter_tools.');
-    // TODO(peterdjlee): Implement a way to unzip the APK for Windows. See issue #62603.
-    String unzipOut;
-    try {
-      // TODO(peterdjlee): Use zipinfo instead of unzip.
-      unzipOut = (await processUtils.run(<String>[
-        'unzip',
-        '-o',
-        '-v',
-        apk.path,
-        '-d',
-        tempApkContent.path
-      ])).stdout;
-    } on Exception catch (e) {
-      logger.printError(e.toString());
-    } finally {
-      // We just want the the stdout printout. We don't need the files.
-      tempApkContent.deleteSync(recursive: true);
-    }
-
-    final _SymbolNode apkAnalysisRoot = _parseUnzipFile(unzipOut);
+    final _SymbolNode aotAnalysisJson = _parseDirectory(
+      outputDirectory,
+      outputDirectory.parent.path,
+      excludePath,
+    );
 
     // Convert an AOT snapshot file into a map.
     final Map<String, dynamic> processedAotSnapshotJson = treemapFromJson(
@@ -75,66 +54,122 @@ class SizeAnalyzer {
     );
     final _SymbolNode aotSnapshotJsonRoot = _parseAotSnapshot(processedAotSnapshotJson);
 
-    for (final _SymbolNode firstLevelPath in apkAnalysisRoot.children) {
+    for (final _SymbolNode firstLevelPath in aotAnalysisJson.children) {
       _printEntitySize(
         firstLevelPath.name,
         byteSize: firstLevelPath.byteSize,
         level: 1,
       );
       // Print the expansion of lib directory to show more info for `appFilename`.
-      if (firstLevelPath.name == 'lib') {
-        _printLibChildrenPaths(firstLevelPath, '', aotSnapshotJsonRoot);
+      if (firstLevelPath.name == fileSystem.path.basename(outputDirectory.path)) {
+        _printLibChildrenPaths(firstLevelPath, '', aotSnapshotJsonRoot, _kAotSizeMaxDepth, 0);
       }
     }
 
     logger.printStatus('▒' * tableWidth);
 
+    Map<String, dynamic> apkAnalysisJson = aotAnalysisJson.toJson();
+
+    apkAnalysisJson['type'] = type; // one of apk, aab, ios, macos, windows, or linux.
+
+    apkAnalysisJson = _addAotSnapshotDataToAnalysis(
+      apkAnalysisJson: apkAnalysisJson,
+      path: _locatedAotFilePath,
+      aotSnapshotJson: processedAotSnapshotJson,
+      precompilerTrace: json.decode(precompilerTrace.readAsStringSync()) as Map<String, Object>,
+    );
+
+    assert(_appFilename != null);
+    return apkAnalysisJson;
+  }
+
+  /// Analyzes [apk] and [aotSnapshot] to output a [Map] object that includes
+  /// the breakdown of the both files, where the breakdown of [aotSnapshot] is placed
+  /// under 'lib/arm64-v8a/$_appFilename'.
+  ///
+  /// [kind] must be one of 'apk' or 'aab'.
+  /// The [aotSnapshot] can be either instruction sizes snapshot or a v8 snapshot.
+  Future<Map<String, dynamic>> analyzeZipSizeAndAotSnapshot({
+    @required File zipFile,
+    @required File aotSnapshot,
+    @required File precompilerTrace,
+    @required String kind,
+  }) async {
+    assert(kind == 'apk' || kind == 'aab');
+    logger.printStatus('▒' * tableWidth);
+    _printEntitySize(
+      '${zipFile.basename} (total compressed)',
+      byteSize: zipFile.lengthSync(),
+      level: 0,
+      showColor: false,
+    );
+    logger.printStatus('━' * tableWidth);
+
+    final _SymbolNode apkAnalysisRoot = _parseUnzipFile(zipFile);
+
+    // Convert an AOT snapshot file into a map.
+    final Map<String, dynamic> processedAotSnapshotJson = treemapFromJson(
+      json.decode(aotSnapshot.readAsStringSync()),
+    );
+    final _SymbolNode aotSnapshotJsonRoot = _parseAotSnapshot(processedAotSnapshotJson);
+    for (final _SymbolNode firstLevelPath in apkAnalysisRoot.children) {
+      _printLibChildrenPaths(firstLevelPath, '', aotSnapshotJsonRoot, _kZipSizeMaxDepth, 0);
+    }
+    logger.printStatus('▒' * tableWidth);
+
     Map<String, dynamic> apkAnalysisJson = apkAnalysisRoot.toJson();
 
-    apkAnalysisJson['type'] = 'apk';
+    apkAnalysisJson['type'] = kind;
 
-    // TODO(peterdjlee): Add aot snapshot for all platforms.
     assert(_appFilename != null);
-    apkAnalysisJson = _addAotSnapshotDataToApkAnalysis(
+    apkAnalysisJson = _addAotSnapshotDataToAnalysis(
       apkAnalysisJson: apkAnalysisJson,
-      path: 'lib/arm64-v8a/$_appFilename (Dart AOT)'.split('/'), // Pass in a list of paths by splitting with '/'.
+      path: _locatedAotFilePath,
       aotSnapshotJson: processedAotSnapshotJson,
+      precompilerTrace: json.decode(precompilerTrace.readAsStringSync()) as Map<String, Object>,
     );
 
     return apkAnalysisJson;
   }
 
-
-  // Expression to match 'Size' column to group 1 and 'Name' column to group 2.
-  final RegExp _parseUnzipOutput = RegExp(r'^\s*\d+\s+[\w|:]+\s+(\d+)\s+.*  (.+)$');
-
-  // Parse the output of unzip -v which shows the zip's contents' compressed sizes.
-  // Example output of unzip -v:
-  //  Length   Method    Size  Cmpr    Date    Time   CRC-32   Name
-  // --------  ------  ------- ---- ---------- ----- --------  ----
-  //    11708  Defl:N     2592  78% 00-00-1980 00:00 07733eef  AndroidManifest.xml
-  //     1399  Defl:N     1092  22% 00-00-1980 00:00 f53d952a  META-INF/CERT.RSA
-  //    46298  Defl:N    14530  69% 00-00-1980 00:00 17df02b8  META-INF/CERT.SF
-  _SymbolNode _parseUnzipFile(String unzipOut) {
+  _SymbolNode _parseUnzipFile(File zipFile) {
+    final Archive archive = ZipDecoder().decodeBytes(zipFile.readAsBytesSync());
     final Map<List<String>, int> pathsToSize = <List<String>, int>{};
 
-    // Parse each path into pathsToSize so that the key is a list of
-    // path parts and the value is the size.
-    // For example:
-    // 'path/to/file' where file = 1500 => pathsToSize[['path', 'to', 'file']] = 1500
-    for (final String line in const LineSplitter().convert(unzipOut)) {
-      final RegExpMatch match = _parseUnzipOutput.firstMatch(line);
-      if (match == null) {
+    for (final ArchiveFile archiveFile in archive.files) {
+      pathsToSize[fileSystem.path.split(archiveFile.name)] = archiveFile.rawContent.length;
+    }
+    return _buildSymbolTree(pathsToSize);
+  }
+
+  _SymbolNode _parseDirectory(Directory directory, String relativeTo, String excludePath) {
+    final Map<List<String>, int> pathsToSize = <List<String>, int>{};
+    for (final File file in directory.listSync(recursive: true).whereType<File>()) {
+      if (excludePath != null && file.uri.pathSegments.contains(excludePath)) {
         continue;
       }
-      const int sizeGroupIndex = 1;
-      const int nameGroupIndex = 2;
-      pathsToSize[match.group(nameGroupIndex).split('/')] = int.parse(match.group(sizeGroupIndex));
+      final List<String> path = fileSystem.path.split(
+        fileSystem.path.relative(file.path, from: relativeTo));
+      pathsToSize[path] = file.lengthSync();
     }
+    return _buildSymbolTree(pathsToSize);
+  }
 
-    final _SymbolNode rootNode = _SymbolNode('Root');
+  List<String> _locatedAotFilePath;
 
+  List<String> _buildNodeName(_SymbolNode start, _SymbolNode parent) {
+    final List<String> results = <String>[start.name];
+    while (parent != null && parent.name != 'Root') {
+      results.insert(0, parent.name);
+      parent = parent.parent;
+    }
+    return results;
+  }
+
+  _SymbolNode _buildSymbolTree(Map<List<String>, int> pathsToSize) {
+     final _SymbolNode rootNode = _SymbolNode('Root');
     _SymbolNode currentNode = rootNode;
+
     for (final List<String> paths in pathsToSize.keys) {
       for (final String path in paths) {
         _SymbolNode childWithPathAsName = currentNode.childByName(path);
@@ -144,6 +179,7 @@ class SizeAnalyzer {
           if (matchesPattern(path, pattern: appFilenamePattern) != null) {
             _appFilename = path;
             childWithPathAsName.name += ' (Dart AOT)';
+            _locatedAotFilePath = _buildNodeName(childWithPathAsName, currentNode);
           } else if (path == 'libflutter.so') {
             childWithPathAsName.name += ' (Flutter Engine)';
           }
@@ -154,7 +190,6 @@ class SizeAnalyzer {
       }
       currentNode = rootNode;
     }
-
     return rootNode;
   }
 
@@ -165,52 +200,79 @@ class SizeAnalyzer {
     _SymbolNode currentNode,
     String totalPath,
     _SymbolNode aotSnapshotJsonRoot,
+    int maxDepth,
+    int currentDepth,
   ) {
     totalPath += currentNode.name;
 
     assert(_appFilename != null);
     if (currentNode.children.isNotEmpty
-      && currentNode.name != '$_appFilename (Dart AOT)') {
+      && currentNode.name != '$_appFilename (Dart AOT)'
+      && currentDepth < maxDepth
+      && currentNode.byteSize >= 1000) {
       for (final _SymbolNode child in currentNode.children) {
-        _printLibChildrenPaths(child, '$totalPath/', aotSnapshotJsonRoot);
+        _printLibChildrenPaths(child, '$totalPath/', aotSnapshotJsonRoot, maxDepth, currentDepth + 1);
       }
+      _leadingPaths = totalPath.split('/')
+        ..removeLast();
     } else {
-      // Print total path and size if currentNode does not have any chilren.
-      _printEntitySize(totalPath, byteSize: currentNode.byteSize, level: 2);
-
-      // We picked this file because arm64-v8a is likely the most popular
-      // architecture. Other architecture sizes should be similar.
-      final String libappPath = 'lib/arm64-v8a/$_appFilename';
-      // TODO(peterdjlee): Analyze aot size for all platforms.
-      if (totalPath.contains(libappPath)) {
-        _printAotSnapshotSummary(aotSnapshotJsonRoot);
+      // Print total path and size if currentNode does not have any children and is
+      // larger than 1KB
+      final bool isAotSnapshotPath = _locatedAotFilePath.join('/').contains(totalPath);
+      if (currentNode.byteSize >= 1000 || isAotSnapshotPath) {
+        _printEntitySize(totalPath, byteSize: currentNode.byteSize, level: 1, emphasis: currentNode.children.isNotEmpty);
+        if (isAotSnapshotPath) {
+          _printAotSnapshotSummary(aotSnapshotJsonRoot, level: totalPath.split('/').length);
+        }
+        _leadingPaths = totalPath.split('/')
+          ..removeLast();
       }
     }
   }
 
   /// Go through the AOT gen snapshot size JSON and print out a collapsed summary
   /// for the first package level.
-  void _printAotSnapshotSummary(_SymbolNode aotSnapshotRoot, {int maxDirectoriesShown = 10}) {
+  void _printAotSnapshotSummary(_SymbolNode aotSnapshotRoot, {int maxDirectoriesShown = 20, @required int level}) {
     _printEntitySize(
       'Dart AOT symbols accounted decompressed size',
       byteSize: aotSnapshotRoot.byteSize,
-      level: 3,
+      level: level,
+      emphasis: true,
     );
 
     final List<_SymbolNode> sortedSymbols = aotSnapshotRoot.children.toList()
+      // Remove entries like  @unknown, @shared, and @stubs as well as private dart libraries
+      //  which are not interpretable by end users.
+      ..removeWhere((_SymbolNode node) => node.name.startsWith('@') || node.name.startsWith('dart:_'))
       ..sort((_SymbolNode a, _SymbolNode b) => b.byteSize.compareTo(a.byteSize));
     for (final _SymbolNode node in sortedSymbols.take(maxDirectoriesShown)) {
-      _printEntitySize(node.name, byteSize: node.byteSize, level: 4);
+      // Node names will have an extra leading `package:*` name, remove it to
+      // avoid extra nesting.
+      _printEntitySize(_formatExtraLeadingPackages(node.name), byteSize: node.byteSize, level: level + 1);
     }
   }
 
+  String _formatExtraLeadingPackages(String name) {
+    if (!name.startsWith('package')) {
+      return name;
+    }
+    final List<String> chunks = name.split('/');
+    if (chunks.length < 2) {
+      return name;
+    }
+    chunks.removeAt(0);
+    return chunks.join('/');
+  }
+
   /// Adds breakdown of aot snapshot data as the children of the node at the given path.
-  Map<String, dynamic> _addAotSnapshotDataToApkAnalysis({
+  Map<String, dynamic> _addAotSnapshotDataToAnalysis({
     @required Map<String, dynamic> apkAnalysisJson,
     @required List<String> path,
     @required Map<String, dynamic> aotSnapshotJson,
+    @required Map<String, dynamic> precompilerTrace,
   }) {
     Map<String, dynamic> currentLevel = apkAnalysisJson;
+    currentLevel['precompiler-trace'] = precompilerTrace;
     while (path.isNotEmpty) {
       final List<Map<String, dynamic>> children = currentLevel['children'] as List<Map<String, dynamic>>;
       final Map<String, dynamic> childWithPathAsName = children.firstWhere(
@@ -223,14 +285,16 @@ class SizeAnalyzer {
     return apkAnalysisJson;
   }
 
+  List<String> _leadingPaths = <String>[];
+
   /// Print an entity's name with its size on the same line.
   void _printEntitySize(
     String entityName, {
     @required int byteSize,
     @required int level,
     bool showColor = true,
+    bool emphasis = false,
   }) {
-    final bool emphasis = level <= 1;
     final String formattedSize = _prettyPrintBytes(byteSize);
 
     TerminalColor color = TerminalColor.green;
@@ -240,12 +304,32 @@ class SizeAnalyzer {
       color = TerminalColor.yellow;
     }
 
-    final int spaceInBetween = tableWidth - level * 2 - entityName.length - formattedSize.length;
+    // Compute any preceeding directories, and compare this to the stored
+    // directoried (in _leadingPaths) for the last entity that was printed. The
+    // similary determines whether or not leading directory information needs to
+    // be printed.
+    final List<String> localSegments = entityName.split('/')
+        ..removeLast();
+    int i = 0;
+    while (i < _leadingPaths.length && i < localSegments.length && _leadingPaths[i] == localSegments[i]) {
+      i += 1;
+    }
+    for (; i < localSegments.length; i += 1) {
+      logger.printStatus(
+        localSegments[i] + '/',
+        indent: (level + i) * 2,
+        emphasis: true,
+      );
+    }
+    _leadingPaths = localSegments;
+
+    final String baseName = fileSystem.path.basename(entityName);
+    final int spaceInBetween = tableWidth - (level + i) * 2 - baseName.length - formattedSize.length;
     logger.printStatus(
-      entityName + ' ' * spaceInBetween,
+      baseName + ' ' * spaceInBetween,
       newline: false,
       emphasis: emphasis,
-      indent: level * 2,
+      indent: (level + i) * 2,
     );
     logger.printStatus(formattedSize, color: showColor ? color : null);
   }
