@@ -521,8 +521,9 @@ class RenderCustomPaint extends RenderProxyBox {
     markNeedsSemanticsUpdate();
   }
 
-  void _paintWithPainter(Canvas canvas, Offset offset, CustomPainter painter) {
-    int debugPreviousCanvasSaveCount;
+  void _paintWithPainter(Canvas canvas, Offset offset, CustomPainter/*!*/ painter) {
+    // TODO(ianh): make the next line final
+    /*late*/ int/*!*/ debugPreviousCanvasSaveCount;
     canvas.save();
     assert(() {
       debugPreviousCanvasSaveCount = canvas.getSaveCount();
@@ -603,10 +604,10 @@ class RenderCustomPaint extends RenderProxyBox {
   }
 
   /// Describe the semantics of the picture painted by the [painter].
-  List<SemanticsNode> _backgroundSemanticsNodes;
+  List<SemanticsNode/*!*/> _backgroundSemanticsNodes;
 
   /// Describe the semantics of the picture painted by the [foregroundPainter].
-  List<SemanticsNode> _foregroundSemanticsNodes;
+  List<SemanticsNode/*!*/> _foregroundSemanticsNodes;
 
   @override
   void assembleSemanticsNode(
@@ -675,7 +676,7 @@ class RenderCustomPaint extends RenderProxyBox {
   /// considered because there is only one type of [SemanticsNode]. There is no
   /// concept of a "forgotten" node in semantics, deactivated nodes, or global
   /// keys.
-  static List<SemanticsNode> _updateSemanticsChildren(
+  static List<SemanticsNode/*!*/> _updateSemanticsChildren(
     List<SemanticsNode> oldSemantics,
     List<CustomPainterSemantics> newChildSemantics,
   ) {
@@ -683,7 +684,7 @@ class RenderCustomPaint extends RenderProxyBox {
     newChildSemantics = newChildSemantics ?? const <CustomPainterSemantics>[];
 
     assert(() {
-      final Map<Key, int> keys = HashMap<Key, int>();
+      final Map<Key/*!*/, int> keys = HashMap<Key/*!*/, int>();
       final List<DiagnosticsNode> information = <DiagnosticsNode>[];
       for (int i = 0; i < newChildSemantics.length; i += 1) {
         final CustomPainterSemantics child = newChildSemantics[i];
@@ -708,7 +709,7 @@ class RenderCustomPaint extends RenderProxyBox {
     int newChildrenBottom = newChildSemantics.length - 1;
     int oldChildrenBottom = oldSemantics.length - 1;
 
-    final List<SemanticsNode> newChildren = List<SemanticsNode>(newChildSemantics.length);
+    final List<SemanticsNode/*!*/> newChildren = List<SemanticsNode/*!*/>(newChildSemantics.length);
 
     // Update the top of the list.
     while ((oldChildrenTop <= oldChildrenBottom) && (newChildrenTop <= newChildrenBottom)) {
@@ -734,7 +735,7 @@ class RenderCustomPaint extends RenderProxyBox {
 
     // Scan the old children in the middle of the list.
     final bool haveOldChildren = oldChildrenTop <= oldChildrenBottom;
-    Map<Key, SemanticsNode> oldKeyedChildren;
+    Map<Key/*!*/, SemanticsNode> oldKeyedChildren;
     if (haveOldChildren) {
       oldKeyedChildren = <Key, SemanticsNode>{};
       while (oldChildrenTop <= oldChildrenBottom) {
