@@ -573,7 +573,7 @@ class _RenderSingleChildViewport extends RenderBox with RenderObjectWithChildMix
 
   bool _shouldClipAtPaintOffset(Offset paintOffset) {
     assert(child != null);
-    return  paintOffset.dx < 0 ||
+    return paintOffset.dx < 0 ||
       paintOffset.dy < 0 ||
       paintOffset.dx + child.size.width > size.width ||
       paintOffset.dy + child.size.height > size.height;
