@@ -8,19 +8,34 @@ import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
+import 'package:flutter/scheduler.dart';
 
 import '../flutter_test_alternative.dart';
 
 typedef HandleEventCallback = void Function(PointerEvent event);
 
-class TestGestureFlutterBinding extends BindingBase with GestureBinding {
+class TestGestureFlutterBinding extends BindingBase with GestureBinding, SchedulerBinding {
   HandleEventCallback callback;
+  FrameCallback frameCallback;
+  Duration frameTime;
 
   @override
   void handleEvent(PointerEvent event, HitTestEntry entry) {
     super.handleEvent(event, entry);
     if (callback != null)
       callback(event);
+  }
+
+  @override
+  Duration get currentSystemFrameTimeStamp {
+    assert(frameTime != null);
+    return frameTime;
+  }
+
+  @override
+  int scheduleFrameCallback(FrameCallback callback, {bool rescheduling = false}) {
+    frameCallback = callback;
+    return 0;
   }
 }
 
@@ -29,6 +44,7 @@ TestGestureFlutterBinding _binding = TestGestureFlutterBinding();
 void ensureTestGestureBinding() {
   _binding ??= TestGestureFlutterBinding();
   assert(GestureBinding.instance != null);
+  assert(SchedulerBinding.instance != null);
 }
 
 void main() {
@@ -298,5 +314,108 @@ void main() {
       expect(events[4].runtimeType, equals(PointerUpEvent));
       expect(events[4].buttons, equals(0));
     }
+  });
+
+  test('Pointer event resampling', () {
+    const ui.PointerDataPacket packet = ui.PointerDataPacket(
+      data: <ui.PointerData>[
+        ui.PointerData(
+            change: ui.PointerChange.add,
+            physicalX: 0.0,
+            timeStamp: Duration(milliseconds: 0),
+        ),
+        ui.PointerData(
+            change: ui.PointerChange.down,
+            physicalX: 0.0,
+            timeStamp: Duration(milliseconds: 1),
+        ),
+        ui.PointerData(
+            change: ui.PointerChange.move,
+            physicalX: 10.0,
+            timeStamp: Duration(milliseconds: 2),
+        ),
+        ui.PointerData(
+            change: ui.PointerChange.move,
+            physicalX: 20.0,
+            timeStamp: Duration(milliseconds: 3),
+        ),
+        ui.PointerData(
+            change: ui.PointerChange.move,
+            physicalX: 30.0,
+            timeStamp: Duration(milliseconds: 4),
+        ),
+        ui.PointerData(
+            change: ui.PointerChange.up,
+            physicalX: 40.0,
+            timeStamp: Duration(milliseconds: 5),
+        ),
+        ui.PointerData(
+            change: ui.PointerChange.remove,
+            physicalX: 40.0,
+            timeStamp: Duration(milliseconds: 6),
+        ),
+      ],
+    );
+
+    final List<PointerEvent> pointerRouterEvents = <PointerEvent>[];
+    GestureBinding.instance.pointerRouter.addGlobalRoute(pointerRouterEvents.add);
+
+    final List<PointerEvent> events = <PointerEvent>[];
+    _binding.callback = events.add;
+
+    GestureBinding.instance.resamplingEnabled = true;
+    GestureBinding.instance.samplingOffset = const Duration(microseconds: -5500);
+
+    ui.window.onPointerDataPacket(packet);
+
+    // No pointer events should have been dispatched yet.
+    expect(events.length, 0);
+
+    // Frame callback should have been requested.
+    FrameCallback callback = _binding.frameCallback;
+    _binding.frameCallback = null;
+    expect(callback, isNotNull);
+
+    _binding.frameTime = const Duration(milliseconds: 7);
+    callback(Duration.zero);
+
+    // One pointer event should have been dispatched.
+    expect(events.length, 1);
+    expect(events[0].runtimeType, equals(PointerDownEvent));
+    expect(events[0].timeStamp, _binding.frameTime + GestureBinding.instance.samplingOffset);
+    expect(events[0].position, Offset(5.0 / ui.window.devicePixelRatio, 0.0));
+
+    // Second frame callback should have been requested.
+    callback = _binding.frameCallback;
+    _binding.frameCallback = null;
+    expect(callback, isNotNull);
+
+    _binding.frameTime = const Duration(milliseconds: 9);
+    callback(Duration.zero);
+
+    // Second pointer event should have been dispatched.
+    expect(events.length, 2);
+    expect(events[1].timeStamp, _binding.frameTime + GestureBinding.instance.samplingOffset);
+    expect(events[1].runtimeType, equals(PointerMoveEvent));
+    expect(events[1].position, Offset(25.0 / ui.window.devicePixelRatio, 0.0));
+    expect(events[1].delta, Offset(20.0 / ui.window.devicePixelRatio, 0.0));
+
+    // Third frame callback should have been requested.
+    callback = _binding.frameCallback;
+    _binding.frameCallback = null;
+    expect(callback, isNotNull);
+
+    _binding.frameTime = const Duration(milliseconds: 11);
+    callback(Duration.zero);
+
+    // Third pointer event should have been dispatched.
+    expect(events.length, 3);
+    expect(events[2].timeStamp, _binding.frameTime + GestureBinding.instance.samplingOffset);
+    expect(events[2].runtimeType, equals(PointerUpEvent));
+    expect(events[2].position, Offset(40.0 / ui.window.devicePixelRatio, 0.0));
+
+    // No frame callback should have been requested.
+    callback = _binding.frameCallback;
+    expect(callback, isNull);
   });
 }
