@@ -220,7 +220,7 @@ void main() {
                 );
                 children.add(sliver);
                 return SliverPadding(
-                  padding: const EdgeInsets.all(22.0),
+                  padding: const EdgeInsets.only(top: 22.0, bottom: 23.0),
                   sliver: sliver,
                 );
               }),
@@ -234,10 +234,16 @@ void main() {
 
     final RenderObject target = tester.renderObject(find.byWidget(children[5], skipOffstage: false));
     RevealedOffset revealed = viewport.getOffsetToReveal(target, 0.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22);
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22);
 
     revealed = viewport.getOffsetToReveal(target, 1.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22 - 100);
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22 - 100);
+
+    // With rect specified.
+    revealed = viewport.getOffsetToReveal(target, 0.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22 + 2);
+    revealed = viewport.getOffsetToReveal(target, 1.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22 - (200 - 4));
   });
 
   testWidgets('Viewport getOffsetToReveal Sliver - right', (WidgetTester tester) async {
@@ -261,7 +267,7 @@ void main() {
                 );
                 children.add(sliver);
                 return SliverPadding(
-                  padding: const EdgeInsets.all(22.0),
+                  padding: const EdgeInsets.only(left: 22.0, right: 23.0),
                   sliver: sliver,
                 );
               }),
@@ -275,10 +281,16 @@ void main() {
 
     final RenderObject target = tester.renderObject(find.byWidget(children[5], skipOffstage: false));
     RevealedOffset revealed = viewport.getOffsetToReveal(target, 0.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22);
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22);
 
     revealed = viewport.getOffsetToReveal(target, 1.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22 - 100);
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22 - 100);
+
+    // With rect specified.
+    revealed = viewport.getOffsetToReveal(target, 0.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22 + 1);
+    revealed = viewport.getOffsetToReveal(target, 1.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 22 - (200 - 3));
   });
 
   testWidgets('Viewport getOffsetToReveal Sliver - up', (WidgetTester tester) async {
@@ -302,7 +314,7 @@ void main() {
                 );
                 children.add(sliver);
                 return SliverPadding(
-                  padding: const EdgeInsets.all(22.0),
+                  padding: const EdgeInsets.only(top: 22.0, bottom: 23.0),
                   sliver: sliver,
                 );
               }),
@@ -316,17 +328,25 @@ void main() {
 
     final RenderObject target = tester.renderObject(find.byWidget(children[5], skipOffstage: false));
     RevealedOffset revealed = viewport.getOffsetToReveal(target, 0.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22);
+    // Does not include the bottom padding of children[5] thus + 23 instead of + 22.
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 23);
 
     revealed = viewport.getOffsetToReveal(target, 1.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22 - 100);
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 23 - 100);
+
+    // With rect specified.
+    revealed = viewport.getOffsetToReveal(target, 0.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 23 + (100 - 4));
+    revealed = viewport.getOffsetToReveal(target, 1.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, - 200 + 6 * (100 + 22 + 23) - 22 - 2);
   });
 
   testWidgets('Viewport getOffsetToReveal Sliver - up - reverse growth', (WidgetTester tester) async {
     const Key centerKey = ValueKey<String>('center');
+    const EdgeInsets padding = EdgeInsets.only(top: 22.0, bottom: 23.0);
     final Widget centerSliver = SliverPadding(
       key: centerKey,
-      padding: const EdgeInsets.all(22.0),
+      padding: padding,
       sliver: SliverToBoxAdapter(
         child: Container(
           height: 100.0,
@@ -339,7 +359,7 @@ void main() {
       child: const Text('Tile lower'),
     );
     final Widget lowerSliver = SliverPadding(
-      padding: const EdgeInsets.all(22.0),
+      padding: padding,
       sliver: SliverToBoxAdapter(
         child: lowerItem,
       ),
@@ -370,13 +390,20 @@ void main() {
 
     revealed = viewport.getOffsetToReveal(target, 1.0);
     expect(revealed.offset, - 100 - 22 - 100);
+
+    // With rect specified.
+    revealed = viewport.getOffsetToReveal(target, 0.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, - 22 - 4);
+    revealed = viewport.getOffsetToReveal(target, 1.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, -200 - 22 - 2);
   });
 
   testWidgets('Viewport getOffsetToReveal Sliver - left - reverse growth', (WidgetTester tester) async {
     const Key centerKey = ValueKey<String>('center');
+    const EdgeInsets padding = EdgeInsets.only(left: 22.0, right: 23.0);
     final Widget centerSliver = SliverPadding(
       key: centerKey,
-      padding: const EdgeInsets.all(22.0),
+      padding: padding,
       sliver: SliverToBoxAdapter(
         child: Container(
           width: 100.0,
@@ -389,7 +416,7 @@ void main() {
       child: const Text('Tile lower'),
     );
     final Widget lowerSliver = SliverPadding(
-      padding: const EdgeInsets.all(22.0),
+      padding: padding,
       sliver: SliverToBoxAdapter(
         child: lowerItem,
       ),
@@ -421,6 +448,12 @@ void main() {
 
     revealed = viewport.getOffsetToReveal(target, 1.0);
     expect(revealed.offset, - 100 - 22 - 200);
+
+    // With rect specified.
+    revealed = viewport.getOffsetToReveal(target, 0.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, - 22 - 3);
+    revealed = viewport.getOffsetToReveal(target, 1.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, - 300 - 22 - 1);
   });
 
   testWidgets('Viewport getOffsetToReveal Sliver - left', (WidgetTester tester) async {
@@ -445,7 +478,7 @@ void main() {
                 );
                 children.add(sliver);
                 return SliverPadding(
-                  padding: const EdgeInsets.all(22.0),
+                  padding: const EdgeInsets.only(left: 22.0, right: 23.0),
                   sliver: sliver,
                 );
               }),
@@ -459,10 +492,16 @@ void main() {
 
     final RenderObject target = tester.renderObject(find.byWidget(children[5], skipOffstage: false));
     RevealedOffset revealed = viewport.getOffsetToReveal(target, 0.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22);
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 23);
 
     revealed = viewport.getOffsetToReveal(target, 1.0);
-    expect(revealed.offset, 5 * (100 + 22 + 22) + 22 - 100);
+    expect(revealed.offset, 5 * (100 + 22 + 23) + 23 - 100);
+
+    // With rect specified.
+    revealed = viewport.getOffsetToReveal(target, 0.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, 6 * (100 + 22 + 23) - 22 - 3);
+    revealed = viewport.getOffsetToReveal(target, 1.0, rect: const Rect.fromLTRB(1, 2, 3, 4));
+    expect(revealed.offset, -200  + 6 * (100 + 22 + 23) - 22 - 1);
   });
 
   testWidgets('Nested Viewports showOnScreen', (WidgetTester tester) async {
