@@ -4,6 +4,7 @@
 
 import 'dart:async';
 
+import '../base/analyze_size.dart';
 import '../base/common.dart';
 import '../build_info.dart';
 import '../cache.dart';
@@ -30,6 +31,7 @@ class BuildLinuxCommand extends BuildSubCommand {
     addBuildPerformanceFile(hide: !verboseHelp);
     addBundleSkSLPathOption(hide: !verboseHelp);
     addNullSafetyModeOptions(hide: !verboseHelp);
+    usesAnalyzeSizeFlag();
   }
 
   @override
@@ -56,7 +58,15 @@ class BuildLinuxCommand extends BuildSubCommand {
     if (!globals.platform.isLinux) {
       throwToolExit('"build linux" only supported on Linux hosts.');
     }
-    await buildLinux(flutterProject.linux, buildInfo, target: targetFile);
+    await buildLinux(
+      flutterProject.linux,
+      buildInfo,
+      target: targetFile,
+      sizeAnalyzer: SizeAnalyzer(
+        fileSystem: globals.fs,
+        logger: globals.logger,
+      ),
+    );
     return FlutterCommandResult.success();
   }
 }
