@@ -15,10 +15,8 @@ import 'package:mockito/mockito.dart';
 import '../../src/common.dart';
 import '../../src/context.dart';
 
-class MockPlatform extends Mock implements Platform {}
-
 void main() {
-  group('ensureDirectoryExists', () {
+  group('fsUtils', () {
     MemoryFileSystem fs;
     FileSystemUtils fsUtils;
 
@@ -26,18 +24,36 @@ void main() {
       fs = MemoryFileSystem();
       fsUtils = FileSystemUtils(
         fileSystem: fs,
-        platform: MockPlatform(),
+        platform: FakePlatform(),
       );
     });
 
-    testWithoutContext('recursively creates a directory if it does not exist', () async {
+    testWithoutContext('ensureDirectoryExists recursively creates a directory if it does not exist', () async {
       fsUtils.ensureDirectoryExists('foo/bar/baz.flx');
       expect(fs.isDirectorySync('foo/bar'), true);
     });
 
-    testWithoutContext('throws tool exit on failure to create', () async {
+    testWithoutContext('ensureDirectoryExists throws tool exit on failure to create', () async {
       fs.file('foo').createSync();
       expect(() => fsUtils.ensureDirectoryExists('foo/bar.flx'), throwsToolExit());
+    });
+
+    testWithoutContext('getUniqueFile creates a unique file name', () async {
+      final File fileA = fsUtils.getUniqueFile(fs.currentDirectory, 'foo', 'json')
+        ..createSync();
+      final File fileB = fsUtils.getUniqueFile(fs.currentDirectory, 'foo', 'json');
+
+      expect(fileA.path, '/foo_01.json');
+      expect(fileB.path, '/foo_02.json');
+    });
+
+    testWithoutContext('getUniqueDirectory creates a unique directory name', () async {
+      final Directory directoryA = fsUtils.getUniqueDirectory(fs.currentDirectory, 'foo')
+        ..createSync();
+      final Directory directoryB = fsUtils.getUniqueDirectory(fs.currentDirectory, 'foo');
+
+      expect(directoryA.path, '/foo_01');
+      expect(directoryB.path, '/foo_02');
     });
   });
 
@@ -61,7 +77,7 @@ void main() {
 
       final FileSystemUtils fsUtils = FileSystemUtils(
         fileSystem: sourceMemoryFs,
-        platform: MockPlatform(),
+        platform: FakePlatform(),
       );
       fsUtils.copyDirectorySync(sourceDirectory, targetDirectory);
 
@@ -81,7 +97,7 @@ void main() {
       final MemoryFileSystem fileSystem = MemoryFileSystem();
       final FileSystemUtils fsUtils = FileSystemUtils(
         fileSystem: fileSystem,
-        platform: MockPlatform(),
+        platform: FakePlatform(),
       );
       final Directory origin = fileSystem.directory('/origin');
       origin.createSync();
